+++ src/main/java/rrcf/general/RCLeaf.java
@@ -9,8 +9,8 @@ import java.io.Serializable;
 public class RCLeaf extends RCNode implements Serializable {
     public int depth;
 
-    public RCLeaf(double[] p, int d) {
-        point = new double[1][p.length];
+    public RCLeaf(float[] p, int d) {
+        point = new float[1][p.length];
         point[0] = p;
         depth = d;
         num = 1;
